Add program_exec() and program_spawn() to Python's os module.
Both functions take a file descriptor as their first argument and a Python data
structure as the second. They will convert the Python data structure into
argdata, analogously to how argdata is converted into a Python data structure
when Python starts. Like in the C library, program_exec() replaces the current
process, while program_spawn() creates a child process and returns the file
descriptor.

For file descriptor passing to work, file descriptors must be wrapped in an
object that has a fileno() method, otherwise they are passed as ints. For
sockets and files, this already works; for directories and other kinds of file
descriptors it needs to be done manually.